+++ src/mechanisms/xoauth2/mechinfo.rs
@@ -1,6 +1,6 @@
 use crate::alloc::boxed::Box;
 use crate::mechname::Mechname;
-use crate::registry::{Matches, Mechanism, Name, Side};
+use crate::registry::{Matches, Mechanism, Named, Side};
 
 #[cfg(feature = "registry_static")]
 use crate::registry::{distributed_slice, MECHANISMS};
@@ -14,7 +14,7 @@ use super::{client, server};
 pub static XOAUTH2: Mechanism = Mechanism {
     mechanism: Mechname::const_new(b"XOAUTH2"),
     priority: 300,
-    client: Some(|_sasl| Ok(Box::new(client::XOAuth2::default()))),
+    client: Some(|| Ok(Box::new(client::XOAuth2::default()))),
     server: Some(|_sasl| Ok(Box::new(server::XOAuth2::default()))),
     first: Side::Client,
 
@@ -23,7 +23,7 @@ pub static XOAUTH2: Mechanism = Mechanism {
 };
 
 struct Select;
-impl Name for Select {
+impl Named for Select {
     fn mech() -> &'static Mechanism {
         &XOAUTH2
     }