+++ src/test/mechanism.rs
@@ -5,7 +5,7 @@
 
 use crate::mechanism::{Authentication, MechanismData};
 use crate::prelude::*;
-use crate::registry::{Matches, Mechanism, Name, Side};
+use crate::registry::{Matches, Mechanism, Named, Side};
 use acid_io::Write;
 
 /// X-RSASLTEST implementation
@@ -75,9 +75,7 @@ pub struct RSaslTest {
 }
 impl RSaslTest {
     #[allow(clippy::unnecessary_wraps)]
-    pub fn client(
-        _config: &SASLConfig,
-    ) -> Result<Box<dyn Authentication>, SASLError> {
+    pub fn client() -> Result<Box<dyn Authentication>, SASLError> {
         Ok(Box::new(Self {
             state: RsaslState::New,
         }))
@@ -122,7 +120,7 @@ pub static RSASLTEST_CF: Mechanism = Mechanism {
 };
 
 struct SelectCF;
-impl Name for SelectCF {
+impl Named for SelectCF {
     fn mech() -> &'static Mechanism {
         &RSASLTEST_CF
     }
@@ -139,7 +137,7 @@ pub static RSASLTEST_SF: Mechanism = Mechanism {
 };
 
 struct Select;
-impl Name for Select {
+impl Named for Select {
     fn mech() -> &'static Mechanism {
         &RSASLTEST_SF
     }